pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dmsa_pred"

# version = "1.1.4"

description = ""
readme = "README.md"
authors = [
    { name = "Jared Galloway", email= "jgallowa@fredhutch.org" },
    { name = "Hugh Haddox", email = "hhaddox@fredhutch.org" }
]
#license = { file = "LICENSE.txt" }
#classifiers = [
#    "License :: OSI Approved :: GNU Affero General Public License v3",
#    "Programming Language :: Python",
#    "Programming Language :: Python :: 3",
#]
keywords = ["dmsa_pred"]
dependencies = [
    "click",
    "numpy",
    "pandas",
    "biopython",
    "polyclonal @ git+https://github.com/jbloomlab/polyclonal@fd87e2647f9101f881d6bb25c0d29942253f9d01",
    "augur"
]
requires-python = ">=3.6"

#[project.optional-dependencies]
#dev = [
#    "pytest",
#    "pytest-pep8",
#    "pytest-datadir",
#    "pre-commit",
#    "black",
#    "sphinx",
#    "sphinx_rtd_theme",
#    "sphinx-click",
#    "bumpver"
#]

[tool.poetry]
readme = "README.md"

[project.urls]
#documentation = "https://matsengrp.github.io/dmsa_pred/"
repository = "https://github.com/matsengrp/dmsa_pred"

[tool.setuptools]
packages = ["dmsa_pred"]

[project.scripts]
dmsa_pred = "dmsa_pred.cli:cli"

#[tool.bumpver]
#
#current_version = "1.1.4"
#
#version_pattern = "MAJOR.MINOR.PATCH"
#commit_message = "bump version {old_version} -> {new_version}"
#commit = true
#tag = true
#push = false

#[tool.bumpver.file_patterns]
#"pyproject.toml" = [
#    'current_version = "{version}"', 'version = "{version}"'
#]
#"dmsa_pred/__init__.py" = ["{version}"]
#"Dockerfile" = ['\tversion "{version}"']
#"README.md" = [
#    "{version}",
#]

